Make libecl compatible with Python 3

+++ python/tests/ecl_tests/test_kw_function.py
@@ -36,7 +36,7 @@ class KWFunctionTest(EclTest):
 
         kw = Ecl3DKW.create( "REGIONS" , grid , EclDataType.ECL_INT , global_active = True )
         kw.assign( 0 )
-        kw[0:nx*ny/2] = 1
+        kw[0:int(nx*ny/2)] = 1
         kw[5,2,0] = 0
         kw[0,9,0] = 2
